pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = "Вычислитель отличий – программа, которая определяет разницу между двумя структурами данных. Это популярная задача, для решения которой существует множество онлайн-сервисов, например, jsondiff. Подобный механизм, например, используется при выводе тестов или при автоматическом отслеживании изменений в конфигурационных файлах."
authors = ["NevermoreKatana <crymorebch@gmail.com>"]
license = "Proprietary"
readme = "README.md"
packages = [
  { include = "difference_calculator" },
]

[tool.poetry.dependencies]
python = "^3.10"
PyYAML = "^6.0"

[tool.poetry.group.dev.dependencies]
flake8 = "^6.0.0"
pytest = "^7.4.0"
codeclimate-test-reporter = "^0.2.3"
pytest-cov = "^4.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
cmd_gendiff = "difference_calculator.cmd_gendiff:cmd_gendiff"
gendiff = "difference_calculator.gendiff.generate_diff:main"
parser = "difference_calculator.parser.parser:main"
entry = "difference_calculator.gendiff.gendiff:main"